+++ Adapters/AppLovin/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## AppLovin iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationAppLovin', '13.3.1.0'
+pod 'CASMediationAppLovin', '13.3.1.1'
 ```
+
+### 13.3.1.1
+- The AppLovin dependency is no longer required for other CAS adapters.
 
 ### 13.3.1.0
 - Certified with AppLovin - 13.3.1