+++ Adapters/Ogury/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## Ogury iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationOgury', '5.0.2.0'
+pod 'CASMediationOgury', '5.0.2.1'
 ```
+
+### 5.0.2.1
+- The AppLovin dependency is no longer required for other CAS adapters.
 
 ### 5.0.2.0
 - Certified with Ogury - 5.0.2